refactor: Event-Status에 대한 상수 및 as const 들이 분리되어 있는거 popup-status에 합침
