+++ src/features/host-center/WrapperHostCenter.tsx
@@ -7,18 +7,19 @@ import useGetUserInfo from '@/hooks/useGetUserInfo';
 import type { Popup } from '@/types/popup';
 
 interface WrapperHostCenterProps {
-  hostEvents: Popup[];
+  hostPopups: Popup[];
 }
 
-export default function WrapperHostCenter({ hostEvents }: WrapperHostCenterProps) {
+export default function WrapperHostCenter({ hostPopups }: WrapperHostCenterProps) {
   const { userInfo } = useGetUserInfo();
-  console.log(hostEvents);
+  // TODO: 렌더링 최적화 필요
+  // console.log(hostPopups);
 
   return (
     <main className="min-h-screen">
       <HeroSection title="팝업의 여정 호스트 센터" description="이벤트를 주최하고 관리할 수 있습니다." />
       <HostProfile userInfo={userInfo} />
-      <HostPopupPanel hostEvents={hostEvents} />
+      <HostPopupPanel hostPopups={hostPopups} />
     </main>
   );
 }
